// ProcMgt/ClientApp/src/app/models/groupcompany.ts
export class Groupcompany {
  groupCompanyID: string;
  GroupCompanyName: string;
  GroupCompanyCode: string;
  IsActive: boolean;
  GroupCompanyLogoID: string;
  GroupUploadFileName: string;
  GroupCompanyAddressLine1: string;
  GroupCompanyAddressLine2: string;
  GroupCompanyAddressLine3: string;
  GroupCompanyAddressLine4: string;
  GcompanyTelephoneNo: number;
  GcompanyFax: string;
  GcompanyEmail: string;
  GcompanyWeb: string;
  GcompanyRegistrationNo: string;
  VatRegistrationNo: string;
}